64 bits processing starting with these commit
git-svn-id: https://svn.code.sf.net/p/eq10q/code/trunk@109 8f919145-41b5-4f04-b112-5f015527c22b

+++ dsp/filter.c
@@ -26,7 +26,7 @@
 Filter *FilterInit(double rate)
 {
   Filter *filter = (Filter *)malloc(sizeof(Filter));
-  filter->fs=(float)rate;
+  filter->fs= rate;
   filter->gain = 1.0f;
   filter->freq = 100.0f;
   filter->q = 1.0f;
