+++ dynamics.c
@@ -233,7 +233,7 @@ static void runDyn(LV2_Handle instance, uint32_t sample_count)
   const float dc = exp(-1.0f/(decay * sample_rate * 0.001f)); //Decay constant
 
   float gain_reduction = 0.0f;
-  float input_filtered = 0.0f;
+  double input_filtered = 0.0;
   float input_preL;
   #if NUM_CHANNELS == 2
   float input_preR;
@@ -252,10 +252,10 @@ static void runDyn(LV2_Handle instance, uint32_t sample_count)
 
     //Input gain
     input_preL =  plugin_data->input[0][i] * InputGain;
-    input_filtered = input_preL;
+    input_filtered = (double)input_preL;
     #if NUM_CHANNELS == 2
     input_preR = plugin_data->input[1][i] * InputGain;
-    input_filtered = (input_preL + input_preR)*0.5f;
+    input_filtered = (double)(input_preL + input_preR)*0.5f;
     #endif
     
     //Sample to Input Vumeter
